# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mnemos API",
  "description": "Governed long-term memory and continuous learning for AI agents.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/v1/events": {
   "post": {
    "tags": [
     "events"
    ],
    "summary": "Create Events",
    "operationId": "create_events_v1_events_post",
    "parameters": [
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EventsIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "events"
    ],
    "summary": "List Events",
    "operationId": "list_events_v1_events_get",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Workspace Id"
      }
     },
     {
      "name": "session_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session Id"
      }
     },
     {
      "name": "task_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Task Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Events V1 Events Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/working-memory": {
   "put": {
    "tags": [
     "events"
    ],
    "summary": "Put Working Memory",
    "operationId": "put_working_memory_v1_working_memory_put",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WorkingMemoryIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WorkingMemoryOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "events"
    ],
    "summary": "List Working Memory",
    "operationId": "list_working_memory_v1_working_memory_get",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Workspace Id"
      }
     },
     {
      "name": "agent_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Agent Id"
      }
     },
     {
      "name": "session_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/WorkingMemoryOut"
         },
         "title": "Response List Working Memory V1 Working Memory Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/working-memory/{wm_id}": {
   "delete": {
    "tags": [
     "events"
    ],
    "summary": "Delete Working Memory",
    "operationId": "delete_working_memory_v1_working_memory__wm_id__delete",
    "parameters": [
     {
      "name": "wm_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Wm Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EventItem": {
    "properties": {
     "type": {
      "$ref": "#/components/schemas/EventType"
     },
     "payload": {
      "additionalProperties": true,
      "type": "object",
      "title": "Payload"
     },
     "task_id": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Task Id"
     },
     "occurred_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Occurred At"
     }
    },
    "type": "object",
    "required": [
     "type"
    ],
    "title": "EventItem"
   },
   "EventType": {
    "type": "string",
    "enum": [
     "user",
     "assistant",
     "tool",
     "error",
     "feedback",
     "task",
     "system"
    ],
    "title": "EventType"
   },
   "EventsIn": {
    "properties": {
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "project_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Name"
     },
     "agent_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Name"
     },
     "events": {
      "items": {
       "$ref": "#/components/schemas/EventItem"
      },
      "type": "array",
      "maxItems": 500,
      "minItems": 1,
      "title": "Events"
     }
    },
    "type": "object",
    "required": [
     "workspace_id",
     "events"
    ],
    "title": "EventsIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "WorkingMemoryIn": {
    "properties": {
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "key": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Key"
     },
     "content": {
      "type": "string",
      "maxLength": 20000,
      "minLength": 1,
      "title": "Content"
     },
     "ttl_seconds": {
      "type": "integer",
      "maximum": 2592000.0,
      "minimum": 1.0,
      "title": "Ttl Seconds",
      "default": 3600
     },
     "importance": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Importance"
     }
    },
    "type": "object",
    "required": [
     "workspace_id",
     "key",
     "content",
     "importance"
    ],
    "title": "WorkingMemoryIn"
   },
   "WorkingMemoryOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "key": {
      "type": "string",
      "title": "Key"
     },
     "content": {
      "type": "string",
      "title": "Content"
     },
     "importance": {
      "type": "number",
      "title": "Importance"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "workspace_id",
     "agent_id",
     "session_id",
     "key",
     "content",
     "importance",
     "expires_at",
     "updated_at"
    ],
    "title": "WorkingMemoryOut"
   },
...
  }
 }
}